refactor(posts): move post read access to posts/access.ts with switch dispatch
