+++ packages/common/src/services/posts/index.ts
@@ -9,5 +9,5 @@ export * from './createPostInOrganization';
 export * from './createPostOnProfile';
 export * from './getProfilePosts';
 export * from './listProfilePosts';
-export * from './postReadAccess';
+export * from './access';
 export * from './postContext';
